Split account in smaller components
